Add files via upload
this folder contains the PNG images of the paper Mechanum cards. I think these cards are too complicated for anything we're going to do for the first sprint, with the exception of the basic move, piston fists, and maybe arcane slippers. The cards I have in mind are going to be much simpler.

A lot of the art is "borrowed" from various things, and should probably be replaced for the final product. I also have the .psd files on my computer, so these are all very editable.